app/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.43.0
django-commando==0.1.3
django-crum==0.7.7
idna==2.10
livereload==2.6.3
oauthlib==3.1.0
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
urllib3==1.25.10
